Clippy-clean workspace + [workspace.lints] gate
cargo clippy --fix for the mechanical tier (redundant clones/casts/
formats), then by hand: type aliases for the eight type_complexity sites
(BoundKwargs, DynMethodFn, ProcFn, FrozenKey, TranscodeFallback,
ParsedMembers, BuiltinMethodFn reuse), Caps::is_empty,
ValueSubclass::new -> alloc (returns RObj, not Self), File.write-at-offset
gets an explicit truncate(false), read_int loop over enumerate, ffi
same-type cast dropped. Doc repairs: deleted an orphaned doc block in
call.rs whose function no longer exists, moved the parts()-entry doc onto
lower_string_part where it belongs, retired the stale 'nothing constructs
Dynamic yet' claim (runtime_meta::dynamic_from_proc does). Every member
opts into [workspace.lints]; levels stay warn locally, CI enforces -D.

+++ crates/zeo-rt/src/builtins/stat.rs
@@ -122,13 +122,13 @@ builtin_methods! {
 
     "size" => fn st_size(recv, args, _blk) {
         arity!(args, 0);
-        Ok(RubyValue::Int(recv_stat(recv)?.st.st_size as i64))
+        Ok(RubyValue::Int(recv_stat(recv)?.st.st_size))
     }
     // `size?` is nil for an empty file (the "is there content" predicate).
     "size?" => fn st_size_p(recv, args, _blk) {
         arity!(args, 0);
         let n = recv_stat(recv)?.st.st_size;
-        Ok(if n > 0 { RubyValue::Int(n as i64) } else { RubyValue::Nil })
+        Ok(if n > 0 { RubyValue::Int(n) } else { RubyValue::Nil })
     }
     "zero?" => fn st_zero_p(recv, args, _blk) {
         arity!(args, 0);
@@ -188,7 +188,7 @@ builtin_methods! {
     }
     "blocks" => fn st_blocks(recv, args, _blk) {
         arity!(args, 0);
-        Ok(RubyValue::Int(recv_stat(recv)?.st.st_blocks as i64))
+        Ok(RubyValue::Int(recv_stat(recv)?.st.st_blocks))
     }
     "ftype" => fn st_ftype(recv, args, _blk) {
         arity!(args, 0);
